2007-08-24  Jeff Johnston  <jjohnstn@redhat.com>
        * libm/common/sf_lround.c (lroundf): Cast sizeof calculation
        to int before comparing it to a signed int value.

+++ newlib/libm/common/sf_lround.c
@@ -29,7 +29,7 @@
   w &= 0x7fffff;
   w |= 0x800000;
 
-  if (exponent_less_127 < (8 * sizeof (long int)) - 1)
+  if (exponent_less_127 < (int)((8 * sizeof (long int)) - 1))
     {
       if (exponent_less_127 < 0)
         return exponent_less_127 < -1 ? 0 : sign;
